feat: allowed decimals for fiterate

+++ src/main/java/gg/deepsite/pewpew/api/objects/PewpewWeaponItem.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class PewpewWeaponItem extends PewPewItem {
 
 	private double baseDamage;
-	private int fireRate;
+	private double fireRate;
 	private int reloadTime;
 	private String ammoType;
 	private int maxAmmo;
